+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD (unreleased)
 
+- Do not count trailing if/unless as a kewyword (https://github.com/zombocom/dead_end/pull/44)
+
 ## 1.0.2
 
 - Fix bug where empty lines were interpreted to have a zero indentation (https://github.com/zombocom/dead_end/pull/39)